AsyncTlsConnection: Fix allocation loop (#513)
This class allocates memory in the constructor and frees it when the destructor
is called. If the accept callback is invoked with an error then we create a new
connection object and pass on the existing connection via a shared pointer.
During manual testing we have encountered a scenario in which we keep getting
errors and hence keep creating new connection objects. Eventually, we will run
out of memory. This change replaces the shared pointer so that the previous
connection object can be freed.